requirements.txt
Flask-Login
Flask-SQLAlchemy
Werkzeug
python-dotenv
pandas
matplotlib
pyttsx3
openai-whisper
langchain
langchain-community
langchain-core
google-genai
streamlit-audiorecorder